Fix automatically upgrading to 9.0.11.
I noticed when a new version was posted last weekend that when I restarted my container, it still thinks the latest version is 9.0.9.441. The problem appears to be when sorting, the sort is not version aware. That makes '9.0.9.' comes after '9.0.11' and breaks this way of getting the latest version. I just added -V to sort which makes it version aware and now it selects the right version.

Alternatively the following command could also be used:

```curl -L https://bintray.com/opensagetv/sagetv/SageTV/_latestVersion | egrep -o '\/opensagetv\/sagetv\/SageTV\/9\.[0-9]*\.[0-9]*\.[0-9]*' | tail -1 | sed 's/^.*[^0-9]\([0-9]*\.[0-9]*\.[0-9]*\.[0-9]*\).*$/\1/'```

This is changes the mechanics of the version checking function to use the latest version redirect link that always works on Bintray. I'm listing this as an alternative if for some reason we find another problem with versions being sorted incorrectly.